pyproject.toml
[project]
name = "niffler"
version = "0.1.0"
description = "✨ Niffler is Chasing the Shiny Meme Coins."
authors = [{ name = "binghuis", email = "soongbinghui@gmail.com" }]
dependencies = [
    "uvicorn[standard]>=0.32.1",
    "python-dotenv>=1.0.1",
    "pydantic-settings>=2.6.1",
    "python-telegram-bot>=21.9",
    "httpx>=0.28.1",
    "tweepy>=4.15.0",
    "spacy>=3.8.2",
    "openai>=1.64.0",
    "motor>=3.7.0",
    "beanie>=1.29.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.scripts]
dev = "python pymon.py src/niffler/__main__.py src/niffler"

[dependency-groups]
dev = [
    "watchdog>=6.0.0",
    "rich>=13.9.4",
]
